backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0
alembic==1.12.1
sqlalchemy==2.0.23

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML Libraries
openai==1.3.6
google-generativeai==0.3.2
easyocr==1.7.0
pillow==10.1.0
opencv-python==4.8.1.78

# Video processing (for metadata extraction)
python-ffmpeg==1.0.16
moviepy==1.0.3

# Web scraping and crawling
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0

# File handling
python-magic==0.4.27

# Background tasks
celery[redis]==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and logging
structlog==23.2.0
rich==13.7.0

# API documentation
python-dotenv==1.0.0